Añadido el tipo long en factoriales

+++ Practicas personales/Factoriales.java
@@ -3,7 +3,7 @@ class Factoriales {
 
     // Declaracion de variable numero y darle el numero 1
     int numero = 1;
-    int resultado = 1;
+    long resultado = 1;
     int repeticiones = 8;
     // Bucle para sacar el factorial de 8
     while (numero <= repeticiones) {
